Adding project-root option.
