+++ analysis/config/dataflow_config.go
@@ -109,3 +109,89 @@ type SlicingSpec struct {
 	// bound by a closure. This can be useful to test data flows because bound labels generate a lot of false positives.
 	SkipBoundLabels bool `yaml:"unsafe-skip-bound-labels" json:"unsafe-skip-bound-labels"`
 }
+
+// IsPathSensitiveFunc returns true if funcName matches any regex in c.Options.PathSensitiveFuncs.
+func (c Config) IsPathSensitiveFunc(funcName string) bool {
+	for _, psfr := range c.pathSensitiveFuncsRegexes {
+		if psfr == nil {
+			continue
+		}
+		if psfr.MatchString(funcName) {
+			return true
+		}
+	}
+
+	return false
+}
+
+// Below are functions used to query the configuration on specific facts
+
+func (c Config) isSomeTaintSpecCid(cid CodeIdentifier, f func(t TaintSpec, cid CodeIdentifier) bool) bool {
+	for _, x := range c.DataflowProblems.TaintTrackingProblems {
+		if f(x, cid) {
+			return true
+		}
+	}
+	return false
+}
+
+// IsSomeSource returns true if the code identifier matches any source in the config
+func (c Config) IsSomeSource(cid CodeIdentifier) bool {
+	return c.isSomeTaintSpecCid(cid, func(t TaintSpec, cid2 CodeIdentifier) bool { return t.IsSource(cid2) })
+}
+
+// IsSomeSink returns true if the code identifier matches any sink in the config
+func (c Config) IsSomeSink(cid CodeIdentifier) bool {
+	return c.isSomeTaintSpecCid(cid, func(t TaintSpec, cid2 CodeIdentifier) bool { return t.IsSink(cid2) })
+}
+
+// IsSomeSanitizer returns true if the code identifier matches any sanitizer in the config
+func (c Config) IsSomeSanitizer(cid CodeIdentifier) bool {
+	return c.isSomeTaintSpecCid(cid, func(t TaintSpec, cid2 CodeIdentifier) bool { return t.IsSanitizer(cid2) })
+}
+
+// IsSomeValidator returns true if the code identifier matches any validator in the config
+func (c Config) IsSomeValidator(cid CodeIdentifier) bool {
+	return c.isSomeTaintSpecCid(cid, func(t TaintSpec, cid2 CodeIdentifier) bool { return t.IsValidator(cid2) })
+}
+
+// IsSomeBacktracePoint returns true if the code identifier matches any backtrace point in the slicing problems
+func (c Config) IsSomeBacktracePoint(cid CodeIdentifier) bool {
+	for _, ss := range c.DataflowProblems.SlicingProblems {
+		if ss.IsBacktracePoint(cid) {
+			return true
+		}
+	}
+	return false
+}
+
+// IsSource returns true if the code identifier matches a source specification in the config file
+func (ts TaintSpec) IsSource(cid CodeIdentifier) bool {
+	b := ExistsCid(ts.Sources, cid.equalOnNonEmptyFields)
+	return b
+}
+
+// IsSink returns true if the code identifier matches a sink specification in the config file
+func (ts TaintSpec) IsSink(cid CodeIdentifier) bool {
+	return ExistsCid(ts.Sinks, cid.equalOnNonEmptyFields)
+}
+
+// IsSanitizer returns true if the code identifier matches a sanitizer specification in the config file
+func (ts TaintSpec) IsSanitizer(cid CodeIdentifier) bool {
+	return ExistsCid(ts.Sanitizers, cid.equalOnNonEmptyFields)
+}
+
+// IsValidator returns true if the code identifier matches a validator specification in the config file
+func (ts TaintSpec) IsValidator(cid CodeIdentifier) bool {
+	return ExistsCid(ts.Validators, cid.equalOnNonEmptyFields)
+}
+
+// IsStaticCommand returns true if the code identifier matches a static command specification in the config file
+func (scs StaticCommandsSpec) IsStaticCommand(cid CodeIdentifier) bool {
+	return ExistsCid(scs.StaticCommands, cid.equalOnNonEmptyFields)
+}
+
+// IsBacktracePoint returns true if the code identifier matches a backtrace point according to the SlicingSpec
+func (ss SlicingSpec) IsBacktracePoint(cid CodeIdentifier) bool {
+	return ExistsCid(ss.BacktracePoints, cid.equalOnNonEmptyFields)
+}
