+++ analysis/config/testdata/test_project_root_loading.yaml
@@ -1,0 +1,6 @@
+options:
+  project-root: "../../../../"
+
+dataflow-problems:
+  user-specs:
+    - "analysis/config/testdata/sample-spec.json"
